Implement update site by CMS fuctionality.

+++ src/main/webapp/java/DAO/MenuOptionDAO.java
@@ -4,5 +4,5 @@ public interface MenuOptionDAO {
 
     String getTitleOfMenuOptionById(int id);
 
-    void updateMenuOptionTitleById(int id, String value);
+    void updateMenuOptionTitleById(long id, String value);
 }
